Create KnapsackHappinessOneDimArray.java
Theory: 0/1 Knapsack Problem
The 0/1 Knapsack Problem is a classic Dynamic Programming (DP) problem where you aim to maximize the value of items in a knapsack without exceeding a given weight capacity. The problem has the 0/1 property, meaning you can either include an item entirely in the knapsack or not include it at all.